final-pjt-back/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
curl_cffi==0.11.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
dotenv==0.9.9
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
somepackage==1.2.3
soupsieve==2.7
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.121
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.13
websockets==15.0.1
yfinance==0.2.61
zipp==3.21.0
